refactor(dsv2-lite): split runtime layout
